requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.0
altair==5.3.0
antlr4-python3-runtime==4.8
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
datasets==2.19.1
defusedxml==0.7.1
diffusers==0.29.0
dill==0.3.8
einops==0.3.0
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
gitpython==3.1.43
grpcio==1.62.1
huggingface-hub==0.23.3
imageio>=2.33
imageio-ffmpeg==0.4.2
importlib-metadata==6.11.0
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.0
lazy-loader==0.4
lightning-utilities==0.11.2
linkify-it-py==2.0.2
markdown==3.6
markdown-it-py==2.2.0
matplotlib==3.8.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
omegaconf==2.1.1
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==21.3
pandas==2.2.2
peft==0.10.0
protobuf==4.25.3
psutil
pudb==2019.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycocotools==2.0.7
pydeck==0.8.1b0
pydeprecate==0.3.1
pyparsing==3.1.2
pytorch-lightning==1.5.0
pytz==2024.1
qudida==0.0.4
referencing==0.34.0
regex==2024.4.16
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
share==1.0.4
smmap==5.0.1
streamlit==1.27.2
supervision==0.22.0
taming-transformers==0.0.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.4.0
tifffile==2024.4.18
timm==1.0.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torchmetrics==0.7.3
tqdm==4.66.2
transformers==4.39.3
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ultralytics==8.2.32
ultralytics-thop==0.2.8
urwid==2.6.10
validators==0.28.0
watchdog==4.0.0
wcwidth==0.2.13
werkzeug==3.0.2
xxhash==3.4.1
yapf==0.32.0
yarl==1.9.4